Moved to new interpolate utils function.

+++ src/editor_components/modulation_meter.h
@@ -50,8 +50,8 @@ class ModulationMeter : public Component {
     const SynthSlider* destination_;
 
     double current_value_;
-    double knob_percent_;
-    double mod_percent_;
+    float knob_percent_;
+    float mod_percent_;
 
     PathStrokeType knob_stroke_;
     float full_radius_;
